requirements.txt
APScheduler==3.10.1
ddddocr==1.4.7
Flask==2.3.2
playwright==1.35.0
Requests==2.31.0

django-scheduler~=0.10.1